Struct aws_sdk_cloudtrail::client::fluent_builders::UpdateTrail

pub struct UpdateTrail<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> UpdateTrail<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(self) -> Result<UpdateTrailOutput, SdkError<UpdateTrailError>> where
R::Policy: SmithyRetryPolicy<UpdateTrailInputOperationOutputAlias, UpdateTrailOutput, UpdateTrailError, UpdateTrailInputOperationRetryAlias>,

pub fn name(self, inp: impl Into<String>) -> Self

Specifies the name of the trail or trail ARN. If Name is a trail name, the string must meet the following requirements:

Contain only ASCII letters (a-z, A-Z), numbers (0-9), periods (.), underscores (_), or dashes (-)

Start with a letter or number, and end with a letter or number

Be between 3 and 128 characters

Have no adjacent periods, underscores or dashes. Names like my-_namespace and my--namespace are not valid.

Not be in IP address format (for example, 192.168.5.4)

If Name is a trail ARN, it must be in the following format.

arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail

pub fn set_name(self, input: Option<String>) -> Self

pub fn s3_bucket_name(self, inp: impl Into<String>) -> Self

Specifies the name of the Amazon S3 bucket designated for publishing log files. See Amazon S3 Bucket Naming Requirements.

pub fn set_s3_bucket_name(self, input: Option<String>) -> Self

pub fn s3_key_prefix(self, inp: impl Into<String>) -> Self

Specifies the Amazon S3 key prefix that comes after the name of the bucket you have designated for log file delivery. For more information, see Finding Your CloudTrail Log Files. The maximum length is 200 characters.

pub fn set_s3_key_prefix(self, input: Option<String>) -> Self

pub fn sns_topic_name(self, inp: impl Into<String>) -> Self

Specifies the name of the Amazon SNS topic defined for notification of log file delivery. The maximum length is 256 characters.

pub fn set_sns_topic_name(self, input: Option<String>) -> Self

pub fn include_global_service_events(self, inp: bool) -> Self

Specifies whether the trail is publishing events from global services such as IAM to the log files.

pub fn set_include_global_service_events(self, input: Option<bool>) -> Self

pub fn is_multi_region_trail(self, inp: bool) -> Self

Specifies whether the trail applies only to the current region or to all regions. The default is false. If the trail exists only in the current region and this value is set to true, shadow trails (replications of the trail) will be created in the other regions. If the trail exists in all regions and this value is set to false, the trail will remain in the region where it was created, and its shadow trails in other regions will be deleted. As a best practice, consider using trails that log events in all regions.

pub fn set_is_multi_region_trail(self, input: Option<bool>) -> Self

pub fn enable_log_file_validation(self, inp: bool) -> Self

Specifies whether log file validation is enabled. The default is false.

When you disable log file integrity validation, the chain of digest files is broken after one hour. CloudTrail does not create digest files for log files that were delivered during a period in which log file integrity validation was disabled. For example, if you enable log file integrity validation at noon on January 1, disable it at noon on January 2, and re-enable it at noon on January 10, digest files will not be created for the log files delivered from noon on January 2 to noon on January 10. The same applies whenever you stop CloudTrail logging or delete a trail.

pub fn set_enable_log_file_validation(self, input: Option<bool>) -> Self

pub fn cloud_watch_logs_log_group_arn(self, inp: impl Into<String>) -> Self

Specifies a log group name using an Amazon Resource Name (ARN), a unique identifier that represents the log group to which CloudTrail logs are delivered. Not required unless you specify CloudWatchLogsRoleArn.

pub fn set_cloud_watch_logs_log_group_arn(self, input: Option<String>) -> Self

pub fn cloud_watch_logs_role_arn(self, inp: impl Into<String>) -> Self

Specifies the role for the CloudWatch Logs endpoint to assume to write to a user's log group.

pub fn set_cloud_watch_logs_role_arn(self, input: Option<String>) -> Self

pub fn kms_key_id(self, inp: impl Into<String>) -> Self

Specifies the KMS key ID to use to encrypt the logs delivered by CloudTrail. The value can be an alias name prefixed by "alias/", a fully specified ARN to an alias, a fully specified ARN to a key, or a globally unique identifier.

CloudTrail also supports KMS multi-Region keys. For more information about multi-Region keys, see Using multi-Region keys in the Key Management Service Developer Guide.

Examples:

alias/MyAliasName

arn:aws:kms:us-east-2:123456789012:alias/MyAliasName

arn:aws:kms:us-east-2:123456789012:key/12345678-1234-1234-1234-123456789012

12345678-1234-1234-1234-123456789012

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn is_organization_trail(self, inp: bool) -> Self

Specifies whether the trail is applied to all accounts in an organization in Organizations, or only for the current Amazon Web Services account. The default is false, and cannot be true unless the call is made on behalf of an Amazon Web Services account that is the management account for an organization in Organizations. If the trail is not an organization trail and this is set to true, the trail will be created in all Amazon Web Services accounts that belong to the organization. If the trail is an organization trail and this is set to false, the trail will remain in the current Amazon Web Services account but be deleted from all member accounts in the organization.

pub fn set_is_organization_trail(self, input: Option<bool>) -> Self