Module aws_sdk_cloudtrail::error

Modules

cloud_trail_access_not_enabled_exception

See CloudTrailAccessNotEnabledException

cloud_trail_arn_invalid_exception

See CloudTrailArnInvalidException

cloud_trail_invalid_client_token_id_exception

See CloudTrailInvalidClientTokenIdException

cloud_watch_logs_delivery_unavailable_exception

See CloudWatchLogsDeliveryUnavailableException

conflict_exception

See ConflictException

insight_not_enabled_exception

See InsightNotEnabledException

insufficient_dependency_service_access_permission_exception

See InsufficientDependencyServiceAccessPermissionException

insufficient_encryption_policy_exception

See InsufficientEncryptionPolicyException

insufficient_s3_bucket_policy_exception

See InsufficientS3BucketPolicyException

insufficient_sns_topic_policy_exception

See InsufficientSnsTopicPolicyException

invalid_cloud_watch_logs_log_group_arn_exception

See InvalidCloudWatchLogsLogGroupArnException

invalid_cloud_watch_logs_role_arn_exception

See InvalidCloudWatchLogsRoleArnException

invalid_event_category_exception

See InvalidEventCategoryException

invalid_event_selectors_exception

See InvalidEventSelectorsException

invalid_home_region_exception

See InvalidHomeRegionException

invalid_insight_selectors_exception

See InvalidInsightSelectorsException

invalid_kms_key_id_exception

See InvalidKmsKeyIdException

invalid_lookup_attributes_exception

See InvalidLookupAttributesException

invalid_max_results_exception

See InvalidMaxResultsException

invalid_next_token_exception

See InvalidNextTokenException

invalid_parameter_combination_exception

See InvalidParameterCombinationException

invalid_s3_bucket_name_exception

See InvalidS3BucketNameException

invalid_s3_prefix_exception

See InvalidS3PrefixException

invalid_sns_topic_name_exception

See InvalidSnsTopicNameException

invalid_tag_parameter_exception

See InvalidTagParameterException

invalid_time_range_exception

See InvalidTimeRangeException

invalid_token_exception

See InvalidTokenException

invalid_trail_name_exception

See InvalidTrailNameException

kms_exception

See KmsException

kms_key_disabled_exception

See KmsKeyDisabledException

kms_key_not_found_exception

See KmsKeyNotFoundException

maximum_number_of_trails_exceeded_exception

See MaximumNumberOfTrailsExceededException

not_organization_master_account_exception

See NotOrganizationMasterAccountException

operation_not_permitted_exception

See OperationNotPermittedException

organization_not_in_all_features_mode_exception

See OrganizationNotInAllFeaturesModeException

organizations_not_in_use_exception

See OrganizationsNotInUseException

resource_not_found_exception

See ResourceNotFoundException

resource_type_not_supported_exception

See ResourceTypeNotSupportedException

s3_bucket_does_not_exist_exception

See S3BucketDoesNotExistException

tags_limit_exceeded_exception

See TagsLimitExceededException

trail_already_exists_exception

See TrailAlreadyExistsException

trail_not_found_exception

See TrailNotFoundException

trail_not_provided_exception

See TrailNotProvidedException

unsupported_operation_exception

See UnsupportedOperationException

Structs

AddTagsError

CloudTrailAccessNotEnabledException

This exception is thrown when trusted access has not been enabled between CloudTrail and Organizations. For more information,

CloudTrailArnInvalidException

This exception is thrown when an operation is called with a trail ARN that is not valid. The following is the format of a trail ARN.

CloudTrailInvalidClientTokenIdException

This exception is thrown when a call results in the InvalidClientTokenId error code.

CloudWatchLogsDeliveryUnavailableException

Cannot set a CloudWatch Logs delivery for this region.

ConflictException

This exception is thrown when the specified resource is not ready for an operation.

CreateTrailError

DeleteTrailError

DescribeTrailsError

GetEventSelectorsError

GetInsightSelectorsError

GetTrailError

GetTrailStatusError

InsightNotEnabledException

If you run GetInsightSelectors on a trail that does not have Insights events enabled, the operation throws the exception InsightNotEnabledException.

InsufficientDependencyServiceAccessPermissionException

This exception is thrown when the IAM user or role that is used to create the organization trail is lacking one or more required permissions for

InsufficientEncryptionPolicyException

This exception is thrown when the policy on the S3 bucket or KMS key is not sufficient.

InsufficientS3BucketPolicyException

This exception is thrown when the policy on the S3 bucket is not sufficient.

InsufficientSnsTopicPolicyException

This exception is thrown when the policy on the Amazon SNS topic is not sufficient.

InvalidCloudWatchLogsLogGroupArnException

This exception is thrown when the provided CloudWatch Logs log group is not valid.

InvalidCloudWatchLogsRoleArnException

This exception is thrown when the provided role is not valid.

InvalidEventCategoryException

Occurs if an event category that is not valid is specified as a value of EventCategory.

InvalidEventSelectorsException

This exception is thrown when the PutEventSelectors operation is called with a number of event

InvalidHomeRegionException

This exception is thrown when an operation is called on a trail from a region other than the region in which the trail was created.

InvalidInsightSelectorsException

The formatting or syntax of the InsightSelectors JSON statement in your PutInsightSelectors or GetInsightSelectors request

InvalidKmsKeyIdException

This exception is thrown when the KMS key ARN is not valid.

InvalidLookupAttributesException

Occurs when a lookup attribute is specified that is not valid.

InvalidMaxResultsException

This exception is thrown if the limit specified is not valid.

InvalidNextTokenException

A token that is not valid, or a token that was previously used in a request with different parameters. This exception is thrown if the token is not valid.

InvalidParameterCombinationException

This exception is thrown when the combination of parameters provided is not valid.

InvalidS3BucketNameException

This exception is thrown when the provided S3 bucket name is not valid.

InvalidS3PrefixException

This exception is thrown when the provided S3 prefix is not valid.

InvalidSnsTopicNameException

This exception is thrown when the provided SNS topic name is not valid.

InvalidTagParameterException

This exception is thrown when the specified tag key or values are not valid.

InvalidTimeRangeException

Occurs if the timestamp values are not valid. Either the start time occurs after the end time, or the time range is outside the range of possible values.

InvalidTokenException

Reserved for future use.

InvalidTrailNameException

This exception is thrown when the provided trail name is not valid. Trail names must meet the following requirements:

KmsException

This exception is thrown when there is an issue with the specified KMS key and the trail can’t be updated.

KmsKeyDisabledException

This exception is no longer in use.

KmsKeyNotFoundException

This exception is thrown when the KMS key does not exist, when the S3 bucket and the

ListPublicKeysError

ListTagsError

ListTrailsError

LookupEventsError

MaximumNumberOfTrailsExceededException

This exception is thrown when the maximum number of trails is reached.

NotOrganizationMasterAccountException

This exception is thrown when the Amazon Web Services account making the request to create or update an organization trail is not the management account for an

OperationNotPermittedException

This exception is thrown when the requested operation is not permitted.

OrganizationNotInAllFeaturesModeException

This exception is thrown when Organizations is not configured to support all features. All features must be enabled in Organizations to support

OrganizationsNotInUseException

This exception is thrown when the request is made from an Amazon Web Services account that is not a member of an organization.

PutEventSelectorsError

PutInsightSelectorsError

RemoveTagsError

ResourceNotFoundException

This exception is thrown when the specified resource is not found.

ResourceTypeNotSupportedException

This exception is thrown when the specified resource type is not supported by CloudTrail.

S3BucketDoesNotExistException

This exception is thrown when the specified S3 bucket does not exist.

StartLoggingError

StopLoggingError

TagsLimitExceededException

The number of tags per trail has exceeded the permitted amount. Currently, the limit is 50.

TrailAlreadyExistsException

This exception is thrown when the specified trail already exists.

TrailNotFoundException

This exception is thrown when the trail with the given name is not found.

TrailNotProvidedException

This exception is no longer in use.

UnsupportedOperationException

This exception is thrown when the requested operation is not supported.

UpdateTrailError

Enums

AddTagsErrorKind

CreateTrailErrorKind

DeleteTrailErrorKind

DescribeTrailsErrorKind

GetEventSelectorsErrorKind

GetInsightSelectorsErrorKind

GetTrailErrorKind

GetTrailStatusErrorKind

ListPublicKeysErrorKind

ListTagsErrorKind

ListTrailsErrorKind

LookupEventsErrorKind

PutEventSelectorsErrorKind

PutInsightSelectorsErrorKind

RemoveTagsErrorKind

StartLoggingErrorKind

StopLoggingErrorKind

UpdateTrailErrorKind