Module aws_sdk_cloudtrail::input

Modules

add_tags_input

See AddTagsInput

create_trail_input

See CreateTrailInput

delete_trail_input

See DeleteTrailInput

describe_trails_input

See DescribeTrailsInput

get_event_selectors_input

See GetEventSelectorsInput

get_insight_selectors_input

See GetInsightSelectorsInput

get_trail_input

See GetTrailInput

get_trail_status_input

See GetTrailStatusInput

list_public_keys_input

See ListPublicKeysInput

list_tags_input

See ListTagsInput

list_trails_input

See ListTrailsInput

lookup_events_input

See LookupEventsInput

put_event_selectors_input

See PutEventSelectorsInput

put_insight_selectors_input

See PutInsightSelectorsInput

remove_tags_input

See RemoveTagsInput

start_logging_input

See StartLoggingInput

stop_logging_input

See StopLoggingInput

update_trail_input

See UpdateTrailInput

Structs

AddTagsInput

Specifies the tags to add to a trail.

CreateTrailInput

Specifies the settings for each trail.

DeleteTrailInput

The request that specifies the name of a trail to delete.

DescribeTrailsInput

Returns information about the trail.

GetEventSelectorsInput

GetInsightSelectorsInput

GetTrailInput

GetTrailStatusInput

The name of a trail about which you want the current status.

ListPublicKeysInput

Requests the public keys for a specified time range.

ListTagsInput

Specifies a list of trail tags to return.

ListTrailsInput

LookupEventsInput

Contains a request for LookupEvents.

PutEventSelectorsInput

PutInsightSelectorsInput

RemoveTagsInput

Specifies the tags to remove from a trail.

StartLoggingInput

The request to CloudTrail to start logging Amazon Web Services API calls for an account.

StopLoggingInput

Passes the request to CloudTrail to stop logging Amazon Web Services API calls for the specified account.

UpdateTrailInput

Specifies settings to update for the trail.