Module aws_sdk_cloudtrail::model

Modules

advanced_event_selector

See AdvancedEventSelector

advanced_field_selector

See AdvancedFieldSelector

data_resource

See DataResource

event

See Event

event_selector

See EventSelector

insight_selector

See InsightSelector

lookup_attribute

See LookupAttribute

public_key

See PublicKey

resource

See Resource

resource_tag

See ResourceTag

tag

See Tag

trail

See Trail

trail_info

See TrailInfo

Structs

AdvancedEventSelector

Advanced event selectors let you create fine-grained selectors for the following CloudTrail

AdvancedFieldSelector

A single selector statement in an advanced event selector.

DataResource

The Amazon S3 buckets, Lambda functions, or Amazon DynamoDB tables that you specify

Event

Contains information about an event that was returned by a lookup request. The result includes a representation of a CloudTrail event.

EventSelector

Use event selectors to further specify the management and data event settings for your trail. By

InsightSelector

A JSON string that contains a list of insight types that are logged on a trail.

LookupAttribute

Specifies an attribute and value that filter the events returned.

PublicKey

Contains information about a returned public key.

Resource

Specifies the type and name of a resource referenced by an event.

ResourceTag

A resource tag.

Tag

A custom key-value pair associated with a resource such as a CloudTrail trail.

Trail

The settings for a trail.

TrailInfo

Information about a CloudTrail trail, including the trail's name, home region, and Amazon Resource Name (ARN).

Enums

EventCategory

InsightType

LookupAttributeKey

ReadWriteType