Module aws_sdk_cloudtrail::operation

Structs

AddTags

Adds one or more tags to a trail, up to a limit of 50. Overwrites an existing tag's value when a new value is specified for an existing tag key.

CreateTrail

Creates a trail that specifies the settings for delivery of log data to an Amazon S3 bucket.

DeleteTrail

Deletes a trail. This operation must be called from the region in which the trail was

DescribeTrails

Retrieves settings for one or more trails associated with the current region for your account.

GetEventSelectors

Describes the settings for the event selectors that you configured for your trail.

GetInsightSelectors

Describes the settings for the Insights event selectors that you configured for your trail. GetInsightSelectors shows

GetTrail

Returns settings information for a specified trail.

GetTrailStatus

Returns a JSON-formatted list of information about the specified trail. Fields include information on delivery errors, Amazon SNS and Amazon S3 errors, and start and stop logging times for each trail. This operation returns trail status from a single region. To return trail status from all regions, you must call the operation on each region.

ListPublicKeys

Returns all public keys whose private keys were used to sign the digest files within the specified time range. The public key is needed to validate digest files that were signed with its corresponding private key.

ListTags

Lists the tags for the trail in the current region.

ListTrails

Lists trails that are in the current account.

LookupEvents

Looks up management events or

PutEventSelectors

Configures an event selector or advanced event selectors for your trail.

PutInsightSelectors

Lets you enable Insights event logging by specifying the Insights

RemoveTags

Removes the specified tags from a trail.

StartLogging

Starts the recording of Amazon Web Services API calls and log file delivery for a trail. For a trail that is enabled in all regions, this operation must be called from the region in which the trail was created. This operation cannot be called on the shadow trails (replicated trails in other regions) of a trail that is enabled in all regions.

StopLogging

Suspends the recording of Amazon Web Services API calls and log file delivery for the specified trail.

UpdateTrail

Updates trail settings that control what events you are logging, and how to handle log files. Changes to a trail do not require