Module aws_sdk_cloudtrail::output

Modules

add_tags_output

See AddTagsOutput

create_trail_output

See CreateTrailOutput

delete_trail_output

See DeleteTrailOutput

describe_trails_output

See DescribeTrailsOutput

get_event_selectors_output

See GetEventSelectorsOutput

get_insight_selectors_output

See GetInsightSelectorsOutput

get_trail_output

See GetTrailOutput

get_trail_status_output

See GetTrailStatusOutput

list_public_keys_output

See ListPublicKeysOutput

list_tags_output

See ListTagsOutput

list_trails_output

See ListTrailsOutput

lookup_events_output

See LookupEventsOutput

put_event_selectors_output

See PutEventSelectorsOutput

put_insight_selectors_output

See PutInsightSelectorsOutput

remove_tags_output

See RemoveTagsOutput

start_logging_output

See StartLoggingOutput

stop_logging_output

See StopLoggingOutput

update_trail_output

See UpdateTrailOutput

Structs

AddTagsOutput

Returns the objects or data if successful. Otherwise, returns an error.

CreateTrailOutput

Returns the objects or data listed below if successful. Otherwise, returns an error.

DeleteTrailOutput

Returns the objects or data listed below if successful. Otherwise, returns an error.

DescribeTrailsOutput

Returns the objects or data listed below if successful. Otherwise, returns an error.

GetEventSelectorsOutput

GetInsightSelectorsOutput

GetTrailOutput

GetTrailStatusOutput

Returns the objects or data listed below if successful. Otherwise, returns an error.

ListPublicKeysOutput

Returns the objects or data listed below if successful. Otherwise, returns an error.

ListTagsOutput

Returns the objects or data listed below if successful. Otherwise, returns an error.

ListTrailsOutput

LookupEventsOutput

Contains a response to a LookupEvents action.

PutEventSelectorsOutput

PutInsightSelectorsOutput

RemoveTagsOutput

Returns the objects or data listed below if successful. Otherwise, returns an error.

StartLoggingOutput

Returns the objects or data listed below if successful. Otherwise, returns an error.

StopLoggingOutput

Returns the objects or data listed below if successful. Otherwise, returns an error.

UpdateTrailOutput

Returns the objects or data listed below if successful. Otherwise, returns an error.